mcp-telegram-client/pyproject.toml
[project]
name = "mcp-telegram-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.2",
]
